requirements.txt
python-telegram-bot
openai
pillow
python-dotenv
requests
